Initial framework for compilation logic.
Note that this is heavily WIP and there are lots of TODOs. The general idea is
that the compile package will be responsible for going from the AST to some
in-memory representation of all types, constants, etc. Where the AST says
"reference to type X", the compile module will actually replace that with a
reference to all knowledge we have about type X or fail with an error.