Generate {Value,MapItem}List aliases for containers.
For every list `[]$foo`, we generate the following to satisfy the
`wire.ValueList` interface:

    type _List_$foo_ValueList []$foo

    func (v _List_$foo_ValueList) ForEach(f func(wire.Value) error) error {
        for _, x := range v {
            if err := f($toWire(x)); err != nil {
                return err
            }
        }
    }

Where `$toWire` is how we convert values of type `$foo` to `Value`s.

Similarly, a `ValueList` is generated for sets (`map[$foo]struct{}`) and a
`MapItemList` for `map[$foo]$bar`.

These are all used for lazy serialization of collections. `FromWire`
implementations have no use of this.