Finish FromWire implementation for everything.
Some weirdness had to be done because structs are special cased to always be
passed by reference, which means that typedefs for structs should do the same.
And the reference vs non-reference thing along with pointers affects how
`FromWire` should work. I've verified these changes by hand but we'll
obviously want tests for all the weird corner cases to get better confidence
in this.

To make fromWire work consistently between types, every custom types gets a
function similar to the map/list/set `_Read` function:

    func _$typeName_Read(v wire.Value) ($typeName, error) {
        // ...
    }

This gives us a consistent return type to work with in different `FromWire`
cases.

----

Also note that this commit is hitting a major milestone finally: the generated
code actually compiles (once you fix the package name). I was able to replace
the hand-written prototype code for the following two structs with generated
code. The benchmark we had written against the hand-written version was easily
converted to use the generated code and ran with similar performance
characteristics as the hand-written version.

    struct NestedMixed {
      1: required list<set<i32>> int_set_list
      2: required map<i32,set<string>> map_int_strset
      3: required list<map<i32,set<string>>> map_int_strset_list
    }

    struct A {
        1: required string name
        2: required i64 birthDay
        3: required string phone
        4: required i32 siblings
        5: required bool spouse
        6: required double money
    }