containers: Fix "use of untyped nil"
When parsing containers with primitives, the `fromWire` expression is,
`value.Get$Type(), nil` (where `$Type` is the primitive type). While reading
an item from Value, this becomes,

    foo, err := value.GetI32(), nil

Assigning nil to a variable with no type information will result in "use of
untyped nil". To fix this, we need to add, `var err error` on the line before
it.

    var err error
    foo, err := value.GetI32(), nil

This changes the code generator for lists, sets, and maps to do this.

Also added tests for containers containing primitive types.