Refactor code generation logic
This cleans up some of the complex conditions in the template and deletes
some functions in favor of cleaner versions.

The only change in actual generated output is that for optional primitive
fields, instead of,

    x, err := f.Value.GetBool(), error(nil)
    v.BoolValue = &x

We'll do,

    var x bool
    x, err = f.Value.GetBool(), error(nil)
    v.BoolValue = &x

That is, the err variable will be picked up from the surrounding scope rather
than being declared anew. This makes the contract for this snippet similar to
that for the other `FromWire` logic:

    v.Contact, err = _ContactInfo_Read(f.Value)
    v.EmailAddress, err = f.Value.GetString(), error(nil)