Make one package per Thrift file; support includes
Generated code for each Thrift file now goes into its own package. If a type
defined in a Thrift file is imported and referenced in another Thrift file, the
generated code for that file will include an import for the original Thrift
file's generated package. If a file is imported but nothing from it is
referenced, an import is *not* added, but code for it is still generated.

The generated package structure is calculated based on the locations of the
Thrift files relative to a 'Thrift root' directory. By default, that's just
the deepest common ancestor of all Thrift files in a dependency tree but it
may be specified explicitly with the `-thrift-root` option. So, if the Thrift
root is foo/ and the Thrift files are foo/a.thrift and foo/common/b.thrift,
the generated packages are foo/a and foo/common/b.

The import paths of these packages are calculated based on the location of the
output directory relative to `$GOPATH/src`. This may be provided explicitly
with the `-pkg-prefix` option. This option MUST be provided if the output
directory is not in the `$GOPATH`.

The following other command line options were added:

- `-out DIR`: The directory into which all the generated code will be written.
  Defaults to the current directory.
- `-no-recurse`: Used to disable code generation for Thrift files `include`d
  by the main Thrift file. Note that this is the opposite of Apache Thrift's
  behavior, which defaults to not generating code for included Thrift files.

Resolves #74