Detect and fail on complex typedef cycles
After the link stage, this will walk all `typedef` reference trees and look
for any cyclic references.

The full reference cycle is printed in the error message. For example, given
the input,

    typedef x foo
    struct x { 1: required y a }
    struct y { 1: optional map<string, z> b }
    struct z { 1: optional foo c }

The error is:

    cannot compile "[..]/tmp.thrift": cannot compile "foo": found a type reference cycle:
        foo
     -> x
     -> y
     -> map<string, z>
     -> z
     -> foo

If the cyclic references are spread across multiple files, the output also
includes the file names.

    cannot compile "[..]/tmp.thrift": cannot compile "foo": found a type reference cycle:
        foo ([..]/tmp.thrift)
     -> x ([..]/tmp.thrift)
     -> y ([..]/tmp.thrift)
     -> map<string, z>
     -> z ([..]/b.thrift)
     -> foo ([..]/tmp.thrift)