Support for unhashable types
This adds support for sets and map keys that use unhashable types. A type is
considered hashable only if it is a non-binary primitive type. All other types
are considered unhashable.

For sets with unhashable items, we simply use a list. For maps with unhashable
keys, we use a list of key-value pairs.