constant values: Support casting
This adds support for constant values to be casted to types based on context.
The following casts are supported:

-   `int` to `double`
-   `int` to `bool`
-   `int` to `enum`
-   `map[string]*` to `struct`
-   `list` to `set`

This is the same behavior supported by Apache Thrift, but with an extra level
of safety and type checking.

We need this before we can implement code generation for constant values.