Support for unhashable types (#118)
This adds support for using unhashable types as map keys and set items. A type
is considered hashable only if it is a non-`binary` primitive type. All other
types are considered unhashable.

Sets of unhashable objects are downgraded to lists, and maps with unhashable
keys are downgraded to lists of key-value pairs. De-duplication is left up to
the user.

Resolves #75.