internal: Add frame reader, writer, client & server (#176)
These allow sending and receiving framed requests over arbitrary
`io.{Reader,Writer}`s. Each request is prefixed with a 4-byte big-endian
integer containing the length of the rest of the frame. This is compatible
with Apache Thrift's `TFramedTransport`. ThriftRW plugins will use framed
requests and responses to talk to the ThriftRW process.