gen/containers: Gracefully handle nil items (#183)
It's incorrect to pass nil as an item in a list, set or map (whether as a key
or as a value). Thrift does not support nil values inside containers.

Previously, we would simply panic if a user tried to do this. (The system
would call `ToWire` on the nill value and panic.) With this change, we instead
send an error back to the user indicating that nil values are not allowed.