plugins: Parse command line arguments
This adds support for parsing plugin arguments over the command line. It does
not actually integrate into the code generator yet.

Plugins are specified over the command line using the "-p" or "--plugin"
options. The value is in the form,

    NAME [ARGS ...]

Where `NAME` is the name of the plugin and `ARGS` is zero or more arguments to
pass to the plugin.

An executable with the name `thriftrw-plugin-${NAME}` is expected to be
available on `$PATH` to call the plugin. `ARGS` are passed to the plugin
executable.

For example,

    -p yarpc -p 'accessors --no-setters'

Will call,

-   `thrifrw-plugin-yarpc`
-   `thriftrw-plugin-accessors --no-setters`

Currently, a plugin can only be called once in a single invocation of
ThriftRW. That is, the following is currently unsupported:

    -p "foo --someflag" -p "foo --anotherflag"