plugin: Don't require unique names
This changes MultiHandle and MultiServiceGenerator to stop keying off of the
plugin name. An advantage of this is that plugin names are no longer required
to be unique so we don't have to worry about builtin plugins conflicting with
user plugins. Additionally, this also makes parameterized plugins possible.
That is, you could now have a plugin which generates significantly different
code based on its command line arguments and have it used twice in the same
thriftrw invocation:

    -p 'myplugin --generateA' -p 'myplugin --generateB'