version: Add thriftrw.disableVersionCheck tag
This adds a `thriftrw.disableVersionCheck` build tag which disables the
version check for generated code when enabled. By default, the version check
is still enabled. To disable version checks, use

    go build -tags=thriftrw.disableVersionCheck

The generated executable will no longer have that check.

We need this so that we can use development versions of thriftrw to generate
internal Thrift code.