internal: Make go generate Just Work™ (#261)
We have a little hack in place in internal/plugin so that we can generate
mocks for an internal interface. That relies on having access to the binary of
mockgen compiled from vendor/. Previously, we expected the caller to install
that binary but that breaks other mockgen calls. This changes the script to
build its own binary for this call.