Embedding IDLs in the generated code (#299)
All IDLs used for code generation are embedded as constant strings. This will
obviously bloat the binary on disk. But this wont take resident RAM.

A new structure must be shared by all generated packages for the
representation of a ThriftModule. A ThriftModule references included
ThriftModules. This means you get the full IDL tree from any entry point.

Motivation: YARPC services use ThriftRW for code generation when using Thrift
IDLs. It came up few times that we need access to the IDL sources used during
the build. Looking up files in a repo is inconvenient nor a guarantee of
correctness. By embedding IDL files in the generated code. And thus in the
binary itself, we guarantee correct introspection.