ast: Add LineNumber function (#314)
This adds an ast.LineNumber function which returns the line on which a
Node was defined. This only works if the Node actually records the line
number it was defined on.

This is a useful function to have for code that is using `ast.Walk`. There
will be need of it in the linter.