main.go: exit non-zero if flag parsing fails (#316)
Currently, if an error is generated while parsing flags, an error
message is emitted by the go-flags package, but we then return a nil
error from do() causing the main function to exit with a zero status.

Let's take over the responsibility of emitting error messages from
go-flags and then propogate any errors that are produced during parsing
back up to the main function.  This means we need to handle ErrHelp
ourselves.