gen: Fix bug for typedefed defaults
We have a bug where if a field with a default value is a typedef of a
primitive type, the generated code will be equivalent to,

    type Timestamp int64

    var value *Timestamp = ptr.Int64(42)

This will fail because you cannot assign a `*int64` to a `*Timestamp`.

This commit fixes this issue by generating a `_Timestamp_ptr` function
similar to one generated for other types.