gen: Include docstrings from IDL in code
This changes code generation to include docstrings from the IDL into the
generated code.

A nice side-effect of this is that our Plugin API package
go.uber.org/thriftrw/plugin/api now has documentation outside the Thrift
file.

We include docstrings for all entities that support docstrings except
services and functions. That's because ThriftRW doesn't generate any
entity for the service or function. We can include this information into
the plugin API in a future change.

Resolves #100