gen: Reproduce comments and formatting (#327)
This changes code generation in gen to reproduce the comments and
formatting from the codegen templates.

As a result of this, the generated code is significantly more readable.
See gen/testdata/constants/constants.go for an example.

This also allows us to include comments in the generated code. We can do
that in a follow up change.