gen: Expose ThriftPacakgeImporter interface (#350)
The thriftPackageImporter interface is unexported but still used by the
exported NewGenerator function, making it completely unusable outside
the gen module.

This change exports the thriftPackageImporter interface, making it
possible to use gen.NewGenerator directly. This API lets users control
the import paths that will be used for Thrift files if they're using
ThriftRW as a library.