gen/enum: Add `go.label` to override item name (#363)
This adds support for a `go.label` annotation on enum items which
overrides the "label" for those items. The label for an enum item is
the string representation of its name in Go code. This is the name
we'll use in `{Marshal, Unmarshal}{Text, JSON}`, `String()`, and
`MarshalLogObject`.

This change adds support for enum items only. Support for the same
label should be added to struct field names (#370).

Resolves #362