gen: Generalize labelling logic
This generalizes the labelling logic (go.label) to support any
NamedEntity. To clarify: This isn't actually adding support for new
entities, only generalizing the logic so that it can be re-used for
struct field labels in the future.

As part of generalization, this logic was moved to its own file.