compile/struct: Check field ID bounds
Thrift field IDs are sent over the wire as 16-bit integers. Previously,
we were silently truncating the values specified in the IDL down to 16
bits.

This changes the compiler to validate field IDs during compilation. In
addition to field IDs that are too large, we reject the following as
well:

- Field ID 0: This is reserved for the return value of a non-void
  method.
- Negative field IDs: This is just wrong

Resolves #72