gen/test/quick: More extensive quick check
This expands our usage of testing/quick to verify round-tripping to more
types and simplifies the code to reduce use of reflection. I realized
that the method signatures we rely on are static so we can just cast to
that interface; apparently we were already doing this elsewhere in the
tests.

I also found an issue in the test itself where lazy lists were not being
evaluated so certain invalid values were failing to roundtrip in that
test. This was fixed by introducing a check against wire.EvaluateValue.

Finally, some types had to be excluded from the quick check because they
leave too many invalid paths for testing/quick to explore.