gen/test/quick: Test Equals methods
This adds a sanity check on the Equals methods for generated types. We
need the ability to skip this check for cases involving maps with
unhashable keys because the randomly generated values for those don't
have a guarantee of unique map keys.

This test exposed a bug in the typedef Equals implementation when the
target type is an enum: the generated code would attempt to call Equals
on the underlying type without casting. This commit fixes that issue as
well.