gen/test/quick: Test for enums
This adds tests for enum values to the quick test. This required adding
support for custom generators to the test because with enums, there's a
small number of values we actually recognize. Picking completely random
values will only exercise the "unknown enum value" case.

This exposed a bug where we weren't generating `*_Values` functions for
empty enums.