envelope/test: Refactor to reduce codegen
In #357, we moved the generated code in gen/testdata to
gen/internal/tests. This made the generated types inaccessible from
envelope/envelope_test, which used one of those types.

In the interest of keeping that change isolated, we opted to regenerate
a copy of that code under envelope/internal/tests. This had the effect
of introducing a large amount of generated code, most of which we don't
need.

This change refactors the envelope_test so that we don't need access to
that generated code and deletes the now unnecessary code.