gen/zap: Don't panic on nil structs
Previously, the generated code for structs was not handling the case
when MarshalLogObject is called on a nil value.

```
var x *Foo
x.MarshalLogObject(enc)
```

This is a problem if we pass Zap a non-nil zapcore.ObjectMarshaler with
a nil value. (See https://tour.golang.org/methods/12.)

This changes the generated code to handle the nil receiver case
explicitly.

Caveat: With this change, when you attempt to log a non-nil
zapcore.ObjectMarshaler with a nil struct, it will log an empty object,
not `null`. This isn't perfect but it's better than panicking. Arguably,
attempting to log a nil value with a non-nil interface is approaching
undefined behavior territory, so graceful degradation is better than
nothing.

Resolves #390