gen/struct: IsSet methods for nillable fields
This changes struct generation to include `IsSet*` methods for nillable
fields of a struct. Note that this applies not only to fields marked
optional, but also to required fields that can be nil: maps, slices,
binary, structs, etc.

This change makes it easier to migrate code from Apache Thrift to
ThriftRW.