plugin: Drop library version matching
This change drops the check made by the ThriftRW Plugin system verifying
that the version of ThriftRW talking to the plugin is the same as the
version of ThriftRW used to compile the plugin.

This check is unnecessay because we already verify the API version,
which is defined in the Thrift file.

This check is more trouble than it's worth as it's constantly tripping
up users and we don't change the plugin API nearly often enough to need
this.