`func <T>Ptr(v <T>) *<T>` for primitive typedefs
Using typedefs on a base type is a common thrift pattern used to
describe the kind of value that is within. While structs are trivial to
form a pointer to (`&Foo{"foo"}`), this is not the case with primitive
types (`type Jack string`, `&Jack("RJ45")`). This adds `<T>Ptr`
functions for primitive type aliases (`JackPtr("RJ45")`).