Add single-file flag to generate a single file output
Adds a flag to generate ThriftRW's output (excluding plugin outputs)
into a single file that is lower-cased named after the value specified
by this flag.

This file is generated in the directory created by ThriftRW, which
is named after the thrift file that thriftRW is invoked on.

A file will be written if a valid value for the single-file flag is
specified.