generateModule: Simplify implementation
With support for only one file, generateModule can be simplified.
It no longer needs to instantiate a map (which receives only one entry).
Instead it can return the intended file path and contents of the file in
one go and let the caller handle merging it into the main files map.