Generate single file for ThriftRW output (#416)
ThriftRW (excluding plugin outputs) now only generates a single file
that is lower-cased named after the value specified
by the added single-file flag. Specifying the OutputFile also
means that code will not be generated for included Thrift Files.

If the single-file flag is not specified, the name of the generated
file will be named after the Thrift file that ThriftRW is invoked upon.

This file is generated in the directory created by ThriftRW, which
is named after the Thrift file that ThriftRW is invoked on.