list: Treat nil as empty for required slices (#428)
Previously, we differentiated between `nil` and empty slices for
required `list` fields, taking the position that if you marked a field
as required, you want to always set it to something. In practice, this
has been a surprising behavior for most users, and a hindrance to
migrating from Apache Thrift. Further, it doesn't align with Go's
position that `nil` is a valid empty slice.

This changes our treatment of required `list` fields to allow nil as
input. Note that the change is made only to required fields, and only
during encoding so that we don't change old behaviors.

Specifically, we continue to differentiate between present and
unspecified in the form of `nil` vs empty for optional fields, and we
continue to fail decoding if the field wasn't actually specified over
the wire.

Ref T4195567