TestDeterminePackagePrefix: Don't rely on GOPATH
If `--package-prefix` is not given, `determinePackagePrefix` is used by
ThriftRW to determine what the import path of a directory is. It does so
by comparing the path to the directory relative to the GOPATH,
accounting for multiple `:`-separated entries being present in `GOPATH`.

Long-term, we should use `go/packages` for this but meanwhile, this test
breaks when ThriftRW itself is not inside GOPATH because the test relies
on the source code for ThriftRW being present inside
`$GOPATH/src/go.uber.org/thriftrw`.

This changes the test to no longer make this assumption, fully managing
the state it needs. This has a nice side-effect of further hermitizing
this test as it no longer relies on system configuration.