make: Fix build and generate targets (#437)
`make generate` depends on the ThriftRW binary being available but the
target to generate that binary doesn't declare any dependencies. This
means that `make generate` could use an outdated binary if available.

This adds all Go files in the repository as dependencies to the ThriftRW
binary so that `make build` and `make generate` use the correct
binaries.