gen: retype Options.Plugin for external implements
Options.Plugin was a internal/plugin.Handle. That meant packages outside
of go.uber.org/thriftrw could not implement a plugin that gen.Generate
could use.

A new, user implementable type, struct is defined: CodeGenerator. It
currently has a single field, the service generator, but it could be
extended by adding new generator fields if we want to maintain API
compatibility (this package currently does not have an API guarantee).
Options.Plugin is now specified as a CodeGenerator.

The previous type included a io.Closer and factory for generators. This
makes more sense from a higher level as plugins are typically forked
programs talking over stdin/out. However, it's a detail the generator
does not care for. It's happy so long as it can call the Generate method
of a ServiceGenerator. Including the I/O detail breaks the abstraction
of what the plugin truly are: a callable that generates code.