Add RootModules to api.GenerateServiceResponse
Even if a Thrift file has no service declaration and therefore no
req.RootServices entries, it's stil useful to have Module information
for the Thrift Files that ThriftRW is invoked with.

To provide an example of where this Module information can be
useful, the Directory field of a Module can help place ThriftRW
plugin generated code for Thrift files that don't contain service
declarations next to where ThriftRW has already generated its code
for that module.