idl: Handle escaped quote characters in strings
We pass string literals through strconv.Unquote(), but we need to
perform a little more preprocessing to handle escape quote characters.
Specifically, we need to "unescape" the "opposite" quote characters to
the kind of string we're about to "unquote". That is, for a single-
quoted strings, we need to unescape \" sequences, and for double-quoted
strings, we unescape \' sequences.