idl: support for auto-assigned field identifiers
Thrift allows implicit assignment of field identifiers. This is legacy
syntax, but it's useful to support it for compatibility with old files.

Auto-assigned fields start at -1 and get increasingly negative. If a
negative field identifier is explicitly given, it will be used as the
new basis for any additional auto-assigned fields. This behavior is
consistent with the Apache Thrift implementation.