fossa: Run separately, only on push
Currently, the FOSSA analysis is set to run as part of CI. This makes it
impossible to run builds for incoming pull requests because PRs don't
have access to the secrets needed for the FOSSA analysis.

Resolve this by running the FOSSA analysis only when we push to a
branch of the repository.