idl: Return structured ParseError from idl.Parse()
Previously, parse errors (idl.internal.parseError) weren't visible to
idl.Parse()'s callers. This was inconvenient because it's useful to have
access to (a) the full list of parse errors and (b) the line number that
caused each parse error.

This change introduces idl.ParseError which provides a list of errors
and their lines.