Implement enveloping for Thrift streaming protocol
Implement "stream.RequestReader" interface method, "ReadRequest", on the Thrift binary protocol. The ReadRequest method reads off the request envelope (if present) from a reader and returns a stream.Reader for the caller to read off the remaining un-enveloped request struct. Additionally, the method returns a stream.ResponseWriter to provide a seamless way for the caller to write the response without having to manage the envelope.

Implement "stream.ResponseWriter" interface method, "WriteResponse", on each of the existing responder types ("NoEnvelopeResponder", "EnvelopeV0Responder, "EnvelopeV1Responder") to handle writing a response envelope with the envelope style of the request.

Implement "stream.Writer" methods WriteEnvelope{Begin,End} to support writing an envelope in a strict, versioned style.