protocol/stream: Define enveloping interfaces (#511)
Defines a "RequestReader" interface type for the protocol/stream
package. The RequestReader interface exposes a single method,
"ReadRequest", to provide Thrift handlers a way to transparently stream
data off the request (stream.Reader) and stream data to the response
(stream.ResponseWriter) while being agnostic to the enveloping style of
a given caller (i.e. no-envelope, strict envelope, non-strict envelope).

Defines a "ResponseWriter" interface type for the protocol/stream
package. The ResponseWriter interface exposes a single method,
WriteResponse, to handle writing the response envelope with the
envelope style of the corresponding request. It returns a stream.Writer
(borrowed) to the caller to write the response bytes.

Exposes new methods on the "stream.Writer" interface,
WriteEnvelope{Begin,End} to support writing an envelope in a strict,
versioned style.