idl: account for tabs in column calculations
Now that we're tracking column position, we need to deal with tabs. This
takes a little extra work because tabs can't simply be translated to
e.g. 8 additional characters; they instead advance the "virtual" column
forward to the next tab stop.

We use a 8-character tab width by default, but this could be made
configurable in the future.