binary: Move Responders and Protocol into package (#516)
In #512 it came up that the `*Responder` types in the protocol package
belong in the protocol/binary package because they're binary-specific.
We found that we cannot move the types into the `binary` package because
`EnvelopeAgnosticProtocol.DecodeRequest` returns a `protocol.Responder`,
which causes an import cycle between protocol and protocol/binary.

This change moves the implementation of the binary protocol, all its
methods, and all responder implementations into a binary/protocol.go
file, including the DecodeRequest method and exposes the raw Protocol
type so that we're not limited by an interface.

To get around the import cycle, we redeclare the Responder interface in
the binary package, and adapt the type from a binary.Protocol into a
protocol.Protocol using an unexported type.

Per apidiff, the net effect of these changes is:

```
--- go.uber.org/thriftrw/protocol/binary ---
Compatible changes:
- Default: added
- EnvelopeV0Responder: added
- EnvelopeV1Responder: added
- NoEnvelopeResponder: added
- Protocol: added
- Responder: added
```

So nothing changes in the top-level `protocol` package (besides some
deprecations) and the `binary` package exports a bunch of new types.

Caveat: I would have preferred not to export the `*Responder` types and
return a `Responder` struct instead of the interface, but unfortunately,
ur hands are tied because of exporting the responder implementations
originally. So `binary.Protocol.ReadRequest` has to be able to return
one of the three Responder types, which is what necessitates the
interface.