stream: Implement Request and Response handling with Enveloping (#526)
Based off of the implementation by @usmyth for enveloping, perform reading out a
request as well as providing an appropriate `stream.ReponseWriter`, based on the
enveloping that was included on the request itself.

In both cases of reading out the request as well as providing a
`stream.ResponseWriter`, the interfaces for these needed to change to support
doing so with enveloping.  Since the envelope's begin and end are meant to
encapsulate the request and response body, an object needs to be provided that
understands how to parse, or write out these bodies respectively.

For the request flow, this object implements a `stream.BodyReader` to parse out
the body and should be a thriftrw object that reflects the request's body.  For
the response flow, a `stream.Enveloper` that reflects the response's body and is
also a compatible thriftrw object, must be provided.

Co-authored-by: Timothy Smyth <smyth@uber.com>
Co-authored-by: Abhinav Gupta <abg@uber.com>