streaming: Skip over unrecognized fields and when field does not match the expected type
In the Wire-based implementation of reading the raw Thrift protocol, the reader
greedily reads over all of the fields in the structs, creating the Wire
intermediary format.  This format is then validated and parsed by the generated
code, populating the appropriate Thrift object with the proper values
associated with the corresponding field ID.

In the stream-based implementation, validation and parsing of the raw Thrift
protocol is done as the bytes are read.  Since unrecognized fields were not
skipped over, this caused issues in future reads as there is now data that
doesn't match the expected Thrift type.

Skip over fields that are unrecognized by the streaming reader and skip fields
that don't match the type expected by the Thrift definition.