string typedef: Fix fmt import and return type bugs
PR (#554) introduced an optimization for the String() method of typedefs.

There are two issues with this change:
- Since the "fmt" import was global, if fmt.Sprint was dropped from
  the typedef code generation the import would be unused.
- Using compiled.RootTypeSpec for isStringType results in
  situations where a non-string type is return from a String() function
  resulting in uncompilable code.
  ```
  func(v StringReDef) String() string {
    x := (stringdef.StringDef)(v)
    return x
  }
  ```
  This happens when a type is a typedef of another string typedef.

This change preserves the demonstrated behavior of (#554) where the String()
method of types that are directly a typedef of a string don't call
fmt.Sprint(x) for their results.

I've added generated code cases to demonsrate and verify the corrected
behavior.