docs: Drop hub command from the RELEASE.md (#585)
The current RELEASE.md relies on the `hub` command. I ran into authorization
issues while trying to use `hub` for the release that were not easy to
resolve. The problems I encountered are captured in a still open `hub`
issue: https://github.com/github/hub/issues/2655. Given the extra steps
required to work around and resolve this issue, I think it would be
simpler to frame the release instructions in terms of local `git`
commands and Github UI.

This change updates the RELEASE.md to replace `hub` commands with local
`git` commands and Github UI instructions.